+++ packages/project-access/test/project/cap.test.ts
@@ -1,4 +1,5 @@
 import { join } from 'path';
+import * as childProcess from 'child_process';
 import * as projectModuleMock from '../../src/project/module-loader';
 import type { Package } from '../../src';
 import { FileName } from '../../src/constants';
@@ -14,6 +15,9 @@ import {
 import { toReferenceUri } from '../../src/project/cap';
 import * as file from '../../src/file';
 import os from 'os';
+
+jest.mock('child_process');
+const childProcessMock = jest.mocked(childProcess, { shallow: true });
 
 describe('Test getCapProjectType()', () => {
     test('Test if valid CAP Node.js project is recognized', async () => {
@@ -318,6 +322,92 @@ describe('Test getCapEnvironment', () => {
         await getCapEnvironment('PROJECT_ROOT');
         expect(forSpy).toHaveBeenCalledWith('cds', 'PROJECT_ROOT');
     });
+
+    test('failed to load cds from any location', async () => {
+        // Mock setup
+        childProcessMock.spawn.mockReturnValueOnce(getChildProcessMock('WRONG'));
+        jest.spyOn(projectModuleMock, 'loadModuleFromProject').mockRejectedValueOnce('ERROR_LOCAL');
+
+        // Test execution
+        try {
+            await getCapEnvironment('PROJECT_ROOT');
+            fail('Call to getCapEnvironment() should have thrown error due to missing cds module but did not');
+        } catch (error) {
+            expect(error.toString()).toContain('ERROR_LOCAL');
+            expect(error.toString()).toContain('@sap/cds-dk');
+        }
+    });
+
+    test('call to cds --version does not contain result', async () => {
+        // Mock setup
+        jest.spyOn(childProcessMock, 'spawn').mockReturnValueOnce(getChildProcessMock(''));
+        const loadSpy = jest.spyOn(projectModuleMock, 'loadModuleFromProject').mockRejectedValueOnce('ERROR_LOCAL');
+
+        // Test execution
+        try {
+            await getCapEnvironment('PROJECT_ROOT');
+            fail('Call to getCapEnvironment() should have thrown error due to missing cds module path but did not');
+        } catch (error) {
+            expect(error.toString()).toContain('Error: Module path not found');
+        }
+        expect(loadSpy).toHaveBeenCalledTimes(1);
+    });
+
+    test('call to cds --version throws error', async () => {
+        // Mock setup
+        jest.spyOn(childProcessMock, 'spawn').mockReturnValueOnce(getChildProcessMock('', true));
+        const loadSpy = jest.spyOn(projectModuleMock, 'loadModuleFromProject').mockRejectedValueOnce('ERROR_LOCAL');
+
+        // Test execution
+        try {
+            await getCapEnvironment('PROJECT_ROOT');
+            fail('Call to getCapEnvironment() should have thrown error due to missing cds module path but did not');
+        } catch (error) {
+            expect(error.toString()).toContain('Error: Module path not found');
+        }
+        expect(loadSpy).toHaveBeenCalledTimes(1);
+    });
+
+    test('with cds loaded from other location than project', async () => {
+        // Mock setup
+        const spawnSpy = jest
+            .spyOn(childProcessMock, 'spawn')
+            .mockReturnValueOnce(getChildProcessMock('anyKey: anyValue\nhome: GLOBAL_ROOT\n'));
+        const forSpy = jest.fn();
+        const loadSpy = jest
+            .spyOn(projectModuleMock, 'loadModuleFromProject')
+            .mockRejectedValueOnce('ERROR_LOCAL')
+            .mockResolvedValueOnce({ default: { env: { for: forSpy } } });
+
+        // Test execution
+        await getCapEnvironment('PROJECT_ROOT');
+
+        // Result check
+        expect(spawnSpy).toBeCalledWith('cds', ['--version'], { cwd: undefined, shell: true });
+        expect(loadSpy).toHaveBeenNthCalledWith(1, 'PROJECT_ROOT', '@sap/cds');
+        expect(loadSpy).toHaveBeenNthCalledWith(2, 'GLOBAL_ROOT', '@sap/cds');
+        expect(forSpy).toBeCalledWith('cds', 'PROJECT_ROOT');
+    });
+
+    function getChildProcessMock(data: any, throwError = false): childProcess.ChildProcess {
+        return {
+            stdout: {
+                on: (type: 'data', cb: (chunk: any) => void) => {
+                    if (type === 'data') {
+                        cb(data);
+                    }
+                }
+            },
+            on: (type: 'close' | 'error', cb: (error?: string) => void) => {
+                if (type === 'close') {
+                    cb();
+                }
+                if (type === 'error' && throwError) {
+                    cb('ERROR');
+                }
+            }
+        } as childProcess.ChildProcess;
+    }
 });
 
 describe('toReferenceUri', () => {